Restore ImageFormat class; remove ImageFormat enum; moving stuff around.
