+++ dali/imgcodec/CMakeLists.txt
@@ -3,7 +3,7 @@ project(dali_imgcodec CUDA CXX C)
 
 set(CMAKE_CUDA_SEPARABLE_COMPILATION ON)
 
-add_subdirectory(format)
+add_subdirectory(formats)
 
 # Get all the source files and dump test files
 collect_headers(DALI_INST_HDRS PARENT_SCOPE)
