+++ dali/imgcodec/formats/bmp.h
@@ -1,3 +1,4 @@
+
 // Copyright (c) 2022, NVIDIA CORPORATION & AFFILIATES. All rights reserved.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
@@ -12,21 +13,21 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#ifndef DALI_IMGCODEC_FORMAT_JPEG_H_
-#define DALI_IMGCODEC_FORMAT_JPEG_H_
+#ifndef DALI_IMGCODEC_FORMATS_BMP_H_
+#define DALI_IMGCODEC_FORMATS_BMP_H_
 
-#include "dali/imgcodec/format/image_format.h"
+#include "dali/imgcodec/image_format.h"
 
 namespace dali {
 namespace imgcodec {
 
-class DLL_PUBLIC JpegParser : public ImageParser {
+class DLL_PUBLIC BmpParser : public ImageParser {
  public:
   ImageInfo GetInfo(EncodedImage *encoded) const override;
   bool CanParse(EncodedImage *encoded) const override;
 };
 
-#endif  // DALI_IMGCODEC_FORMAT_JPEG_H_
+#endif  // DALI_IMGCODEC_FORMATS_BMP_H_
 
 }  // namespace imgcodec
 }  // namespace dali
