+++ dali/imgcodec/formats/webp.h
@@ -1,4 +1,3 @@
-
 // Copyright (c) 2022, NVIDIA CORPORATION & AFFILIATES. All rights reserved.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
@@ -13,21 +12,22 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#ifndef DALI_IMGCODEC_FORMAT_BMP_H_
-#define DALI_IMGCODEC_FORMAT_BMP_H_
+#ifndef DALI_IMGCODEC_FORMATS_WEBP_H_
+#define DALI_IMGCODEC_FORMATS_WEBP_H_
 
-#include "dali/imgcodec/format/image_format.h"
+#include "dali/imgcodec/image_format.h"
 
 namespace dali {
 namespace imgcodec {
 
-class DLL_PUBLIC BmpParser : public ImageParser {
+class DLL_PUBLIC WebpParser : public ImageParser {
  public:
   ImageInfo GetInfo(EncodedImage *encoded) const override;
   bool CanParse(EncodedImage *encoded) const override;
 };
 
-#endif  // DALI_IMGCODEC_FORMAT_BMP_H_
+#endif  // DALI_IMGCODEC_FORMATS_WEBP_H_
+
 
 }  // namespace imgcodec
 }  // namespace dali
